+++ src/components/Post/Post.tsx
@@ -1,7 +1,10 @@
+import List from "components/List/List";
 import Title from "components/Title/Title";
 import { htmlToText } from "html-to-text";
+import { Link } from "react-router-dom";
 
 type postProps = {
+  id: string;
   className?: string;
   access?: string;
   thumbnail?: string;
@@ -15,6 +18,7 @@ type postProps = {
 //TODO: access에 따른 render 여부 접근 여부 설정, like, comment 수로 개별 render
 
 const Post = ({
+  id,
   className,
   access,
   title,
@@ -27,21 +31,27 @@ const Post = ({
   const summaryToText = summary ? htmlToText(summary, { wordwrap: null }) : "";
 
   return (
-    <li className={className}>
-      <div className="post-summary">
-        <Title className="title" level={2} children={title} />
-        <Title className="subTitle" level={3} children={subTitle} />
-        <p className="summary">{summaryToText}</p>
-        <div className="post-info">
-          <span>{date}</span>
-          <span>|</span>
-          <span>좋아요 {likeCount > 99 ? "99+" : likeCount}</span>
-          <span>|</span>
-          <span>레터 {commentCount > 99 ? "99+" : commentCount}</span>
-        </div>
-      </div>
-      <div className="thumbnail" />
-    </li>
+    <List
+      className={className}
+      id={id}
+      children={
+        <Link to={`/${id}`}>
+          <div className="post-summary">
+            <Title className="title" level={2} children={title} />
+            <Title className="subTitle" level={3} children={subTitle} />
+            <p className="summary">{summaryToText}</p>
+            <div className="post-info">
+              <span>{date}</span>
+              <span>|</span>
+              <span>좋아요 {likeCount > 99 ? "99+" : likeCount}</span>
+              <span>|</span>
+              <span>레터 {commentCount > 99 ? "99+" : commentCount}</span>
+            </div>
+          </div>
+          <div className="thumbnail" />
+        </Link>
+      }
+    />
   );
 };
