+++ src/pages/Board.tsx
@@ -1,6 +1,6 @@
 import StyledFooter from "containers/Footer/Footer.styled";
 import StyledHeader from "containers/Header/Header.styled";
-import StyledListContainer from "containers/ListContainer/ListContainer.styled";
+import StyledListContainer from "containers/ListContainer/ListContainer";
 import StyledPagination from "containers/Pagination/Pagination.styled";
 
 const Board = () => {
