+++ src/pages/Welcome.tsx
@@ -2,7 +2,7 @@ import MoreReadButton from "components/Button/MoreReadButton.styled";
 import StyledCarousel from "containers/Carousel/Carousel.styled";
 import StyledFooter from "containers/Footer/Footer.styled";
 import StyledHeader from "containers/Header/Header.styled";
-import StyledListContainer from "containers/ListContainer/ListContainer.styled";
+import StyledListContainer from "containers/ListContainer/ListContainer";
 import { useHistory } from "react-router";
 
 const Welcome = () => {
